requirements.txt
# torch==2.0.0
# torchaudio
# torchvision
# conda install pytorch==2.0.0 torchvision==0.15.0 torchaudio==2.0.0 pytorch-cuda=11.8 -c pytorch -c nvidia
# conda install pytorch==2.4.0 torchvision==0.19.0 torchaudio==2.4.0 pytorch-cuda=12.1 -c pytorch -c nvidia
huggingface-hub
matplotlib==3.7.0
psutil==5.9.4
iopath==0.1.10
pyyaml==6.0
regex==2024.11.6
tokenizers==0.20.3
tqdm
transformers==4.46.1
timm==0.6.13
webdataset==0.2.100
omegaconf==2.3.0
opencv-python==4.7.0.72
decord==0.6.0
sentence-transformers
gradio==3.47.1
accelerate==1.3.0
bitsandbytes==0.37.0
scikit-image
visual-genome
wandb
vllm==0.6.3.post1
peft==0.7.0
librosa # for audio
qwen_vl_utils
git+https://github.com/openai/CLIP.git
open-clip-torch
imageio-ffmpeg
diffusers==0.31.0
pyairports==2.1.1
albumentations==1.4.16
img2dataset